+++ app/javascript/applications/HomebrewsApp/pages/Content/Nimble.jsx
@@ -1,0 +1,58 @@
+import { For, Switch, Match } from 'solid-js';
+
+import { NimbleAncestries, NimbleBooks } from '../../pages';
+import { useAppState, useAppLocale } from '../../context';
+
+const TRANSLATION = {
+  en: {
+    books: 'Books',
+    ancestries: 'Ancestries'
+  },
+  ru: {
+    books: 'Книги',
+    ancestries: 'Расы'
+  },
+  es: {
+    books: 'Libros',
+    ancestries: 'Ancestrías'
+  }
+}
+
+export const Nimble = () => {
+  const [appState, { navigate }] = useAppState();
+
+  const [locale] = useAppLocale();
+
+  return (
+    <>
+      <div class="flex flex-wrap gap-x-4 gap-y-2 my-4">
+        <For each={
+          [
+            'books', 'ancestries'
+          ]
+        }>
+          {(item) =>
+            <p
+              class="homebrew-provider-nav"
+              classList={{ 'active': appState.activePageParams.tab === item }}
+              onClick={() => navigate('nimble', { tab: item })}
+            >{TRANSLATION[locale()][item]}</p>
+          }
+        </For>
+      </div>
+      <Switch fallback={<></>}>
+        <For each={
+          Object.entries({
+            ancestries: NimbleAncestries, books: NimbleBooks
+          })
+        }>
+          {([item, Component]) =>
+            <Match when={appState.activePageParams.tab === item}>
+              <Component />
+            </Match>
+          }
+        </For>
+      </Switch>
+    </>
+  );
+}