+++ app/javascript/applications/HomebrewsApp/pages/ContentPage.jsx
@@ -1,6 +1,6 @@
 import { Switch, Match } from 'solid-js';
 
-import { Daggerheart, Dnd2024 } from '../pages';
+import { Daggerheart, Dnd2024, Nimble } from '../pages';
 
 import { useAppState } from '../context';
 
@@ -15,6 +15,9 @@ export const ContentPage = (props) => {
         </Match>
         <Match when={appState.activePage === 'dnd2024'}>
           <Dnd2024 onNavigate={props.onNavigate} />
+        </Match>
+        <Match when={appState.activePage === 'nimble'}>
+          <Nimble onNavigate={props.onNavigate} />
         </Match>
       </Switch>
     </div>